Release 0.2.0

+++ src/main/kotlin/com/remcoil/config/JwtConfig.kt
@@ -2,5 +2,5 @@ package com.remcoil.config
 
 data class JwtConfig(
     val secret: String,
-    val time: Int
+    val time: Long
 )
